pyproject.toml
[tool.poetry]
name = "nautilus"
version = "0.1.0"
description = ""
authors = ["put101 <tobiaspucher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
nautilus-trader = "^1.182.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
notebook = "^7.0.6"
dask = {extras = ["complete"], version = "^2023.12.0"}
matplotlib = "^3.8.2"
plotly = "^5.18.0"
mplfinance = "^0.12.10b0"
bokeh = "^3.3.2"
mplcursors = "^0.5.2"
dash = "^2.14.2"
pyorbital = "^1.8.1"
hvplot = "^0.9.2"

# put101 = {path = "/Users/tobiaspucher/GitHub/nautilus/put101", develop = true}
datashader = "^0.16.0"
python-statemachine = {extras = ["diagrams"], version = "^2.1.2"}
pydot = "^2.0.0"
influxdb-client = "^1.41.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venv = ".venv"
venvPath = "."

